--> Deprecating OpenAiService and renaming OpenAiServiceV2.
--> Updating the OpenAiController to use updated OpenAiService class.
