+++ sample-code/spring-app/src/main/java/com/sap/ai/sdk/app/services/OpenAiService.java
@@ -3,38 +3,29 @@ package com.sap.ai.sdk.app.services;
 import static com.sap.ai.sdk.foundationmodels.openai.OpenAiModel.GPT_4O;
 import static com.sap.ai.sdk.foundationmodels.openai.OpenAiModel.GPT_4O_MINI;
 import static com.sap.ai.sdk.foundationmodels.openai.OpenAiModel.TEXT_EMBEDDING_3_SMALL;
-import static com.sap.ai.sdk.foundationmodels.openai.model.OpenAiChatCompletionTool.ToolType.FUNCTION;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.JsonMappingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.module.jsonSchema.JsonSchemaGenerator;
 import com.sap.ai.sdk.core.AiCoreService;
+import com.sap.ai.sdk.foundationmodels.openai.OpenAiChatCompletionDelta;
+import com.sap.ai.sdk.foundationmodels.openai.OpenAiChatCompletionRequest;
+import com.sap.ai.sdk.foundationmodels.openai.OpenAiChatCompletionResponse;
 import com.sap.ai.sdk.foundationmodels.openai.OpenAiClient;
-import com.sap.ai.sdk.foundationmodels.openai.model.OpenAiChatCompletionDelta;
-import com.sap.ai.sdk.foundationmodels.openai.model.OpenAiChatCompletionFunction;
-import com.sap.ai.sdk.foundationmodels.openai.model.OpenAiChatCompletionOutput;
-import com.sap.ai.sdk.foundationmodels.openai.model.OpenAiChatCompletionParameters;
-import com.sap.ai.sdk.foundationmodels.openai.model.OpenAiChatCompletionTool;
-import com.sap.ai.sdk.foundationmodels.openai.model.OpenAiChatMessage;
-import com.sap.ai.sdk.foundationmodels.openai.model.OpenAiChatToolCall;
-import com.sap.ai.sdk.foundationmodels.openai.model.OpenAiEmbeddingOutput;
-import com.sap.ai.sdk.foundationmodels.openai.model.OpenAiEmbeddingParameters;
+import com.sap.ai.sdk.foundationmodels.openai.OpenAiEmbeddingRequest;
+import com.sap.ai.sdk.foundationmodels.openai.OpenAiEmbeddingResponse;
+import com.sap.ai.sdk.foundationmodels.openai.OpenAiImageItem;
+import com.sap.ai.sdk.foundationmodels.openai.OpenAiMessage;
+import com.sap.ai.sdk.foundationmodels.openai.OpenAiTool;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
 import java.util.stream.Stream;
 import javax.annotation.Nonnull;
 import lombok.extern.slf4j.Slf4j;
+import lombok.val;
 import org.springframework.stereotype.Service;
 
-/** Service class for OpenAI service */
+/** Service class for OpenAI service using latest convenience api */
 @Service
 @Slf4j
 public class OpenAiService {
-  private static final ObjectMapper JACKSON = new ObjectMapper();
-
   /**
    * Chat request to OpenAI
    *
@@ -42,8 +33,31 @@ public class OpenAiService {
    * @return the assistant message response
    */
   @Nonnull
-  public OpenAiChatCompletionOutput chatCompletion(@Nonnull final String prompt) {
-    return OpenAiClient.forModel(GPT_4O_MINI).chatCompletion(prompt);
+  public OpenAiChatCompletionResponse chatCompletion(@Nonnull final String prompt) {
+    return OpenAiClient.forModel(GPT_4O_MINI)
+        .chatCompletion(new OpenAiChatCompletionRequest(prompt));
+  }
+
+  /**
+   * Chat requests to OpenAI and updating the messages history
+   *
+   * @param previousMessage The request to send to the assistant
+   * @return the assistant message response
+   */
+  @Nonnull
+  public OpenAiChatCompletionResponse messagesHistory(@Nonnull final String previousMessage) {
+    val messagesList = new ArrayList<OpenAiMessage>();
+    messagesList.add(OpenAiMessage.user(previousMessage));
+
+    final OpenAiChatCompletionResponse result =
+        OpenAiClient.forModel(GPT_4O_MINI)
+            .chatCompletion(new OpenAiChatCompletionRequest(messagesList));
+
+    messagesList.add(result.getMessage());
+    messagesList.add(OpenAiMessage.user("What is the typical food there?"));
+
+    return OpenAiClient.forModel(GPT_4O_MINI)
+        .chatCompletion(new OpenAiChatCompletionRequest(messagesList));
   }
 
   /**
@@ -55,9 +69,7 @@ public class OpenAiService {
   @Nonnull
   public Stream<OpenAiChatCompletionDelta> streamChatCompletionDeltas(
       @Nonnull final String message) {
-    final var request =
-        new OpenAiChatCompletionParameters()
-            .addMessages(new OpenAiChatMessage.OpenAiChatUserMessage().addText(message));
+    final var request = new OpenAiChatCompletionRequest(OpenAiMessage.user(message));
 
     return OpenAiClient.forModel(GPT_4O_MINI).streamChatCompletionDeltas(request);
   }
@@ -82,93 +94,50 @@ public class OpenAiService {
    * @return the assistant message response
    */
   @Nonnull
-  public OpenAiChatCompletionOutput chatCompletionImage(@Nonnull final String linkToImage) {
-    final var request =
-        new OpenAiChatCompletionParameters()
-            .addMessages(
-                new OpenAiChatMessage.OpenAiChatUserMessage()
-                    .addText("Describe the following image.")
-                    .addImage(
-                        linkToImage,
-                        OpenAiChatMessage.OpenAiChatUserMessage.ImageDetailLevel.HIGH));
+  public OpenAiChatCompletionResponse chatCompletionImage(@Nonnull final String linkToImage) {
 
-    return OpenAiClient.forModel(GPT_4O).chatCompletion(request);
+    final var userMessage =
+        OpenAiMessage.user("Describe the following image.")
+            .withImage(linkToImage, OpenAiImageItem.DetailLevel.HIGH);
+
+    return OpenAiClient.forModel(GPT_4O)
+        .chatCompletion(new OpenAiChatCompletionRequest(userMessage));
   }
 
   /**
-   * Executes a chat completion request to OpenAI with a tool that calculates the weather.
+   * Chat request to OpenAI with tool that gets the weather for a given location and unit. The tool
+   * executed and the result is sent back to the assistant.
    *
    * @param location The location to get the weather for.
    * @param unit The unit of temperature to use.
    * @return The assistant message response.
    */
   @Nonnull
-  public OpenAiChatCompletionOutput chatCompletionToolExecution(
+  public OpenAiChatCompletionResponse chatCompletionToolExecution(
       @Nonnull final String location, @Nonnull final String unit) {
+    final OpenAiClient client = OpenAiClient.forModel(GPT_4O_MINI);
+
+    final var messages = new ArrayList<OpenAiMessage>();
+    messages.add(OpenAiMessage.user("What's the weather in %s in %s?".formatted(location, unit)));
 
     // 1. Define the function
-    final Map<String, Object> schemaMap = generateSchema(WeatherMethod.Request.class);
-    final var function =
-        new OpenAiChatCompletionFunction()
-            .setName("weather")
-            .setDescription("Get the weather for the given location")
-            .setParameters(schemaMap);
-    final var tool = new OpenAiChatCompletionTool().setType(FUNCTION).setFunction(function);
+    final List<OpenAiTool> tools =
+        List.of(
+            OpenAiTool.forFunction(WeatherMethod::getCurrentWeather)
+                .withArgument(WeatherMethod.Request.class)
+                .withName("weather")
+                .withDescription("Get the weather for the given location"));
 
-    final var messages = new ArrayList<OpenAiChatMessage>();
-    messages.add(
-        new OpenAiChatMessage.OpenAiChatUserMessage()
-            .addText("What's the weather in %s in %s?".formatted(location, unit)));
+    // 2. Assistant calls the function
+    final var request = new OpenAiChatCompletionRequest(messages).withToolsExecutable(tools);
+    final OpenAiChatCompletionResponse response = client.chatCompletion(request);
 
-    // Assistant will call the function
-    final var request =
-        new OpenAiChatCompletionParameters()
-            .addMessages(messages.toArray(OpenAiChatMessage[]::new))
-            .setTools(List.of(tool));
+    // 3. Execute the tool calls
+    messages.add(response.getMessage());
+    messages.addAll(response.executeTools());
 
-    final OpenAiChatCompletionOutput response =
-        OpenAiClient.forModel(GPT_4O_MINI).chatCompletion(request);
-
-    // 2. Optionally, execute the function.
-    final OpenAiChatToolCall toolCall =
-        response.getChoices().get(0).getMessage().getToolCalls().get(0);
-    final WeatherMethod.Request arguments =
-        parseJson(toolCall.getFunction().getArguments(), WeatherMethod.Request.class);
-    final WeatherMethod.Response currentWeather = WeatherMethod.getCurrentWeather(arguments);
-
-    final OpenAiChatMessage.OpenAiChatAssistantMessage assistantMessage =
-        response.getChoices().get(0).getMessage();
-    messages.add(assistantMessage);
-
-    final var toolMessage =
-        new OpenAiChatMessage.OpenAiChatToolMessage()
-            .setToolCallId(toolCall.getId())
-            .setContent(currentWeather.toString());
-    messages.add(toolMessage);
-
-    final var finalRequest =
-        new OpenAiChatCompletionParameters()
-            .addMessages(messages.toArray(OpenAiChatMessage[]::new));
-
-    return OpenAiClient.forModel(GPT_4O_MINI).chatCompletion(finalRequest);
-  }
-
-  private static <T> T parseJson(@Nonnull final String rawJson, @Nonnull final Class<T> clazz) {
-    try {
-      return JACKSON.readValue(rawJson, clazz);
-    } catch (JsonProcessingException e) {
-      throw new IllegalArgumentException("Failed to parse tool call arguments: " + rawJson, e);
-    }
-  }
-
-  private static Map<String, Object> generateSchema(@Nonnull final Class<?> clazz) {
-    final var jsonSchemaGenerator = new JsonSchemaGenerator(JACKSON);
-    try {
-      final var schema = jsonSchemaGenerator.generateSchema(clazz);
-      return JACKSON.convertValue(schema, new TypeReference<>() {});
-    } catch (JsonMappingException e) {
-      throw new IllegalArgumentException("Could not generate schema for " + clazz.getName(), e);
-    }
+    // 4. Have model run the final request with incorporated tool results
+    return client.chatCompletion(request.withMessages(messages));
   }
 
   /**
@@ -178,8 +147,8 @@ public class OpenAiService {
    * @return the embedding response
    */
   @Nonnull
-  public OpenAiEmbeddingOutput embedding(@Nonnull final String input) {
-    final var request = new OpenAiEmbeddingParameters().setInput(input);
+  public OpenAiEmbeddingResponse embedding(@Nonnull final String input) {
+    final var request = new OpenAiEmbeddingRequest(List.of(input));
 
     return OpenAiClient.forModel(TEXT_EMBEDDING_3_SMALL).embedding(request);
   }
@@ -192,12 +161,13 @@ public class OpenAiService {
    * @return the assistant message response
    */
   @Nonnull
-  public OpenAiChatCompletionOutput chatCompletionWithResource(
+  public OpenAiChatCompletionResponse chatCompletionWithResource(
       @Nonnull final String resourceGroup, @Nonnull final String prompt) {
 
     final var destination =
         new AiCoreService().getInferenceDestination(resourceGroup).forModel(GPT_4O);
 
-    return OpenAiClient.withCustomDestination(destination).chatCompletion(prompt);
+    return OpenAiClient.withCustomDestination(destination)
+        .chatCompletion(new OpenAiChatCompletionRequest(prompt));
   }
 }
